Update README.md
Tukey test is a single-step multiple comparison procedure and statistical test. It is a post-hoc analysis, what means that it is used in conjunction with an ANOVA.

It allows to find means of a factor that are significantly different from each other, comparing all possible pairs of means with a t-test like method.